refactor: move `is`, `deepEqual` and `shallowEqual` predicates to `@makeswift/controls`
