+++ .changeset/two-days-repair.md
@@ -1,0 +1,6 @@
+---
+"@makeswift/controls": patch
+"@makeswift/runtime": patch
+---
+
+refactor: move `is`, `deepEqual` and `shallowEqual` predicates to the `@makeswift/controls` package
