+++ packages/controls/src/lib/predicates/index.ts
@@ -1,0 +1,3 @@
+export * from './deepEqual'
+export * from './is'
+export * from './shallowEqual'
