+++ packages/runtime/src/utils/is.ts
@@ -1,5 +1,3 @@
-export default function is(x: unknown, y: unknown): boolean {
-  if (x === y) return x !== 0 || y !== 0 || 1 / x === 1 / y
+import { is } from '@makeswift/controls'
 
-  return x !== x && y !== y
-}
+export default is
